requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
cloudinary==1.25.0
defusedxml==0.7.1
dialogflow-fulfillment==0.4.2
diff-match-patch==20200713
Django==2.2.3
django-bootstrap4==3.0.0
django-import-export==2.5.0
django-jsonfield==1.4.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-preventconcurrentlogins==0.8.2
django-session-timeout==0.1.0
django-smartfields==1.1.3
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
importlib-metadata==2.1.1
jsonfield==3.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.4.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==3.0.0
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
